added discovery page and home about page
