+++ client/components/sidebar.js
@@ -23,7 +23,7 @@ function Sidebar() {
       <ul>
         <li><Link href="/home"><i className="fas fa-home"></i>Home</Link></li>  
         <li><Link href="/search"><i className="fas fa-search"></i>Search</Link></li> 
-        {/*<li><Link href="/discover"><i className="fas fa-search"></i>Discover</Link></li> */}
+        <li><Link href="/discover"><i className="fas fa-search"></i>Discover</Link></li>
         <li><Link href="/searchByIngredient"><i className="fas fa-search"></i>Ingredient Search</Link></li> 
         {currentUser && ( <>
           <li>
@@ -49,4 +49,3 @@ function Sidebar() {
 }
 
 export default Sidebar;
-
